Add a SearchState in addition to HeaderSearchState, only show search controls on pages where it makes sense
